@@ -1,19 +1,18 @@
 plugins {
     id 'com.android.library'
-    id 'com.github.dcendents.android-maven'
 }
 
 group='com.github.bright'
 
 android {
-    compileSdkVersion 28
-    buildToolsVersion '28.0.3'
+    namespace "pl.brightinventions.slf4android"
+    testNamespace "pl.brightinventions.slf4android.androidTest"
+
+    compileSdk 34
 
     defaultConfig {
         minSdkVersion 14
-        targetSdkVersion 28
-        versionCode 14
-        versionName "1.0"
+        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
     }
 
     packagingOptions {
@@ -28,21 +27,20 @@ android {
 }
 
 dependencies {
-    implementation 'com.android.support:support-compat:28.0.0'
-    implementation 'org.slf4j:slf4j-api:1.7.7'
+    implementation 'androidx.core:core:1.12.0'
+    implementation 'org.slf4j:slf4j-api:1.7.36'
     implementation 'com.squareup:seismic:1.0.2'
-    androidTestImplementation 'org.hamcrest:hamcrest-all:1.3'
-    androidTestImplementation('junit:junit:4.11') {
-        exclude module: 'hamcrest-core'
-    }
+    androidTestImplementation 'org.hamcrest:hamcrest-core:1.3'
+    androidTestImplementation 'org.hamcrest:hamcrest-library:1.3'
+    androidTestImplementation('junit:junit:4.13.2')
+    androidTestImplementation 'androidx.test:rules:1.5.0'
     androidTestImplementation 'org.easytesting:fest-util:1.2.5'
 
-    testImplementation 'org.hamcrest:hamcrest-all:1.3'
-    testImplementation('junit:junit:4.11') {
-        exclude module: 'hamcrest-core'
-    }
+    testImplementation 'org.hamcrest:hamcrest-core:1.3'
+    testImplementation 'org.hamcrest:hamcrest-library:1.3'
+    testImplementation('junit:junit:4.13.2')
     testImplementation 'org.easytesting:fest-util:1.2.5'
-    testImplementation('org.robolectric:robolectric:3.0') {
+    testImplementation('org.robolectric:robolectric:4.11.1') {
         exclude module: 'classworlds'
         exclude module: 'commons-logging'
         exclude module: 'httpclient'
@@ -60,5 +58,3 @@ dependencies {
         exclude module: 'wagon-provider-api'
     }
 }
-
-
