@@ -1,7 +1,11 @@
 package pl.brightinventions.slf4android.androidTest;
 
-import android.test.AndroidTestCase;
+import static org.hamcrest.MatcherAssert.assertThat;
+import static org.hamcrest.Matchers.containsString;
+import static org.hamcrest.Matchers.not;
 
+import org.junit.Before;
+import org.junit.Test;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -12,17 +16,12 @@ import java.io.InputStreamReader;
 import pl.brightinventions.slf4android.LogLevel;
 import pl.brightinventions.slf4android.LoggerConfiguration;
 
-import static org.hamcrest.MatcherAssert.assertThat;
-import static org.hamcrest.Matchers.containsString;
-import static org.hamcrest.Matchers.not;
-
-public class LoggerAdapterTests extends AndroidTestCase {
+public class LoggerAdapterTests {
 
     public static final String LOGGER_NAME = "ApplicationTest";
 
-    @Override
-    public void setUp() throws Exception {
-        super.setUp();
+    @Before
+    public void setUp() {
         LoggerConfiguration.resetConfigurationToDefault();
         clearLogcat();
     }
@@ -35,7 +34,8 @@ public class LoggerAdapterTests extends AndroidTestCase {
         }
     }
 
-    public void test_debug_message() throws Exception {
+    @Test
+    public void test_debug_message() {
         getLogger().debug("Hello debug {} {}", 1, 2);
         assertThat(getLastMessage(), containsString("Hello debug 1 2"));
     }
@@ -67,24 +67,28 @@ public class LoggerAdapterTests extends AndroidTestCase {
         return result;
     }
 
-    public void test_debug_message_multi_arg() throws Exception {
+    @Test
+    public void test_debug_message_multi_arg() {
         getLogger().debug("Message {} {} {} {}", 1, 2, 3, 4);
         assertThat(getLastMessage(), containsString("Message 1 2 3 4"));
     }
 
-    public void test_info_message() throws Exception {
+    @Test
+    public void test_info_message() {
         getLogger().info("With string '{}'", "string arg");
         assertThat(getLastMessage(), containsString("With string 'string arg'"));
     }
 
-    public void test_warning_message() throws Exception {
+    @Test
+    public void test_warning_message() {
         getLogger().warn("warning message", new NullPointerException("Bad"));
         String lastMessage = getLastMessage();
         assertThat(lastMessage, containsString("warning message"));
         assertThat(lastMessage, containsString(getClass().getName()));
     }
 
-    public void test_error_message() throws Exception {
+    @Test
+    public void test_error_message() {
         getLogger().error("error message", new IllegalArgumentException("Wrong argument"));
         String lastMessage = getLastMessage();
 
@@ -92,6 +96,7 @@ public class LoggerAdapterTests extends AndroidTestCase {
         assertThat(lastMessage, containsString("IllegalArgumentException"));
     }
 
+    @Test
     public void test_info_message_not_printed_when_level_is_set_to_warning() {
         setLevelTo(LogLevel.WARNING);
         getLogger().info("info message with exception", new NullPointerException("Bad"));
@@ -104,18 +109,21 @@ public class LoggerAdapterTests extends AndroidTestCase {
         configuration.setRootLogLevel(level);
     }
 
+    @Test
     public void test_debug_message_not_printed_when_level_is_set_to_info() {
         setLevelTo(LogLevel.INFO);
         getLogger().debug("debug message with exception", new NullPointerException("Bad"));
         assertThat(getLastMessage(), not(containsString("debug message with exception")));
     }
 
+    @Test
     public void test_trace_message_not_printed_when_level_is_set_to_info() {
         setLevelTo(LogLevel.INFO);
         getLogger().trace("new trace message with exception", new NullPointerException("Bad"));
         assertThat(getLastMessage(), not(containsString("new trace message with exception")));
     }
 
+    @Test
     public void test_trace_message_not_printed_when_level_is_set_to_debug() {
         setLevelTo(LogLevel.DEBUG);
         getLogger().trace("trace message with exception", new NullPointerException("Bad"));
@@ -123,12 +131,14 @@ public class LoggerAdapterTests extends AndroidTestCase {
         assertThat(lastMessage, not(containsString("trace message with exception")));
     }
 
+    @Test
     public void test_error_message_printed_when_level_is_set_to_debug() {
         setLevelTo(LogLevel.DEBUG);
         getLogger().error("error message with exception", new NullPointerException("Bad"));
         assertThat(getLastMessage(), containsString("error message with exception"));
     }
 
+    @Test
     public void test_warning_message_printed_when_level_is_set_to_debug() {
         setLevelTo(LogLevel.DEBUG);
         getLogger().error("new warning message with exception", new NullPointerException("Bad"));
