@@ -1,8 +1,13 @@
 package pl.brightinventions.slf4android.androidTest;
 
-import android.app.Application;
-import android.test.ActivityInstrumentationTestCase2;
+import static androidx.test.platform.app.InstrumentationRegistry.getInstrumentation;
 
+import android.app.Application;
+
+import androidx.test.rule.ActivityTestRule;
+
+import org.junit.Before;
+import org.junit.Test;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -11,7 +16,7 @@ import java.util.concurrent.TimeUnit;
 import pl.brightinventions.slf4android.LoggerConfiguration;
 import pl.brightinventions.slf4android.NotifyDeveloperHandler;
 
-public class NotifyDeveloperHandlerTests extends ActivityInstrumentationTestCase2<TestActivity> {
+public class NotifyDeveloperHandlerTests extends ActivityTestRule<TestActivity> {
 
     private Logger LOG;
     private NotifyDeveloperHandler handler;
@@ -20,22 +25,23 @@ public class NotifyDeveloperHandlerTests extends ActivityInstrumentationTestCase
         super(TestActivity.class);
     }
 
-    @Override
-    protected void setUp() throws Exception {
+    @Before
+    public void setUp() {
         LoggerConfiguration.resetConfigurationToDefault();
         LoggerConfiguration configuration = LoggerConfiguration.configuration();
         Application targetContext = (Application) getInstrumentation().getTargetContext().getApplicationContext();
         handler = configuration.notifyDeveloperHandler(targetContext, "piotr.mionskowski@gmail.com");
         configuration.addHandlerToLogger("", handler);
         LOG = LoggerFactory.getLogger(getClass().getSimpleName());
-        super.setUp();
     }
 
-    public void test_dont_send_message_with_level_lower_than_error() throws Exception {
+    @Test
+    public void test_dont_send_message_with_level_lower_than_error() {
         getActivity();
         LOG.warn("Hello");
     }
 
+    @Test
     public void test_send_message_with_level_error() throws Exception {
         getActivity();
         LOG.warn("Hello");
@@ -43,7 +49,8 @@ public class NotifyDeveloperHandlerTests extends ActivityInstrumentationTestCase
         Thread.sleep(TimeUnit.SECONDS.toMillis(15));
     }
 
-    public void test_send_message_with_custom_subject_body_with_level_error() throws Exception{
+    @Test
+    public void test_send_message_with_custom_subject_body_with_level_error() throws Exception {
         handler.withSubject("Błąd").withBody("Podaj szczegóły błędu: ");
         getActivity();
         LOG.warn("Hello");
