@@ -7,6 +7,6 @@ public class TestActivity extends Activity {
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(pl.brightinventions.slf4android.test.R.layout.test);
+        setContentView(R.layout.test);
     }
 }
